Modify global function dr_or_cr()
This was done by adding constants (ACCOUNT_CLASSES)
and sets (DR_ACCOUNTS, DR_FIRSTS, CR_ACCOUNTS, CR_FIRSTS)
to config.py.

Also made both src/config.py and tests/config.py as close to the same as
possible (confirmed using diff) adding to each an explanation of why
both are needed.
This is in VIOLATION of the SPoT/DRY principle and I'd like to
modify it: perhaps by having a second configuration file to contain only
the DEFAULT_DIR constant (the only part that needs to differ.)