Make methods independent of code schema.
Now that the code schema is defined in the config.py file,
methods have been modified to take advantage of this and not require
rewrites if/when the schema is changed.
It proved impossible to do this for the four Kazan15 entity custom
functions- this fact is reflected in their docstrings.