Prepare for Kazan15 final report(s.)
Modified code to allow zeroing out of income as well as expenses.
Still using the 'split' option of the ChartOfAccounts even though
the decision was made to eliminate the 10th participant from the
calculations.

The kazan_journal (the inputs) and the Kazan15ChartOfAccounts
files have been moved to /var/opt/debk.d/ in order to keep it
out of version control.