Sort out the import hierarchy.
Learned that the Project directory is the one that should be in the
PYTHONPATH but it does not need an __init__.py file in it.
Imports need NOT include the Project directory in their path name but
rather start at directories within the Project directory.

Test suit runs successfully but still need to sort out why in test1.py
the tearDown isn't removing the Manero directory as it should.