Fix Cr/Dr bug.
End of descriptive text was determined by presence of 'Cr' or 'Dr' in
the next line.  This caused things to fail if descriptive text
contained either of these character sequences!  Now the code looks for
' Cr ' or ' Dr '.  This will probably break my efforts to allow
entries to have the account number , the value, and the Cr or Dr in
any order.  We'll probably just enforce the following format:
AcntNum Cr currency